requirements.txt
streamlit
openai
python-dotenv
playsound
pygame